main/requirements.txt
numpy>=1.19.0
pandas>=1.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
seaborn>=0.11.0
umap-learn>=0.5.1